Add the show route for surveys

+++ database/migrations/2023_06_27_081945_create_questions_table.php
@@ -11,7 +11,7 @@ return new class extends Migration {
     {
         Schema::create('questions', function (Blueprint $table) {
             $table->id();
-            $table->unsignedTinyInteger('order');
+            $table->integer('order');
             $table->text('content');
             $table->string('type')->default(QuestionType::SINGLE->value);
             $table->timestamps();
